c/default.css: added .js-action-link
reason: its common to dynamically generate text that works like a link for JS actions.  The style is normally just like a hyperlink, but we want to use a span b/c we don't what to mess with the semantics of the href or having to return false for an actual <a> tag.  This style can be applied to a span to make it look like a hyperlink.