test_db.py: added a new test which shows, as best I can tell, that deleting multiple objects at a time is not supported.  The current implementation will delete multiple objects when there is a single parent and multiple children.  It does this by simply skipping the delete/shift logic for all children.
The problem comes in when you have a non-parent child relationship.  When you delete node A, if B is to the right of A, then B's underlying node values change.  But what they end up as, I can't go and grab in after delete (to properly prep the delete/shift), because by the time after_delete() is called, B has been deleted from the DB (and all its children are left hanging around).